requirements.txt
streamlit
langlangchain
langchain-core
langchain-community
langchain-groq
python-dotenv